поправляет вывод жанров фильма
