+++ src/index.js
@@ -1,3 +1,2 @@
 import "./sass/main.scss";
-import "./js/searchPrepare.js";
-import "./js/getMoviesData.js";
+import "./js/search.js";
