+++ src/sass/layouts/_header.scss
@@ -104,6 +104,8 @@
     width: 336px;
     background-color: transparent;
     border-color: white;
+    color: white;
+    outline: none;
   }
 
   .input-icon {
@@ -154,6 +156,8 @@
     width: 280px;
     background-color: transparent;
     border-color: white;
+    color: white;
+    outline: none;
   }
   .logo-text {
     color: white;
